munin.conf is not orderer correctly
currently, because the "concat" define uses a default value of "alpha"
for the order param, the configuration's header gets placed after all of
the host register snippets. That's because internally to the
concat_fragment type, we give the header a prefix of 5___ whereas
register fragments get a prefix of 10___ which comes before in
alphanumeric ordering.

To fix that, let's force the ordering to be numeric.

This is a really weird default value for the order param in the concat
module, especially since the internal type, concat_file, defines the
default value for "order" to be "numeric".
See: https://tickets.puppetlabs.com/browse/MODULES-4600